requirements.txt
psycopg==3.2.7
alembic==1.15.2
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
Flask==3.1.0
Flask-Migrate==4.0.5
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
gunicorn==23.0.0
itsdangerous==2.2.0
Mako==1.3.10
msgspec==0.19.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
Werkzeug==3.1.3